[wasm64] Fix wasm64 + MINIMAL_RUNTIME
Move the instrumentWasmExportsForMemory64 helper out into its own file
so we have be used under MINIMAL_RUNTIME too.  In the long run this
function should be replaced with something more data driven but for now
this works.

Split out from #16922
